refactor: refactor sound functions
