+++ minesweeper/js/sound.js
@@ -1,5 +1,14 @@
-export { setSound, toggleSound };
+export { setSound, toggleSound, clickSound, markSound, endGameSound, sounds, clickSounds, openPopupSound };
 
+const clickSound = new Audio("./assets/sounds/click-01.mp3");
+const markSound = new Audio("./assets/sounds/click-05.mp3");
+const endGameSound = new Audio("./assets/sounds/end-game-sound.mp3");
+const openPopupSound = new Audio("./assets/sounds/whoosh.mov");
+
+const sounds = [clickSound, markSound, endGameSound, openPopupSound];
+const clickSounds = [clickSound, markSound];
+
+//
 // TOGGLE
 function toggleSound(sounds) {
   const soundButton = document.querySelector(".button__toggle-sound");
@@ -12,6 +21,7 @@ function toggleSound(sounds) {
   localStorage.setItem("volume", volume);
 }
 
+//
 // SET
 function setSound(sounds) {
   const soundButton = document.querySelector(".button__toggle-sound");
